notifications: split classes into respective files
