+++ invenio_users_resources/notifications/filters.py
@@ -6,13 +6,9 @@
 # modify it under the terms of the MIT License; see LICENSE file for more
 # details.
 
-"""User specific resources for notifications."""
+"""User specific filters for notifications."""
 
-
-from invenio_notifications.models import Recipient
-from invenio_notifications.services.builders import RecipientGenerator
 from invenio_notifications.services.filters import RecipientFilter
-from invenio_records.dictutils import dict_lookup
 
 
 class UserPreferencesRecipientFilter(RecipientFilter):
@@ -29,18 +25,4 @@ class UserPreferencesRecipientFilter(RecipientFilter):
             ):
                 del recipients[key]
 
-        return recipients
-
-
-class UserRecipient(RecipientGenerator):
-    """User recipient generator for a notification."""
-
-    def __init__(self, key):
-        """Ctor."""
-        self.key = key
-
-    def __call__(self, notification, recipients):
-        """Update required recipient information and add backend id."""
-        user = dict_lookup(notification.context, self.key)
-        recipients[user["id"]] = Recipient(data=user)
         return recipients
